+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [UNRELEASED] - 2022-##-##
 ### Changed
 - show modal restyled for mobile
+- plex icon sizing
 
 ### Fixed
 - RT/IMDB rating of 0 shown